requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2
groq==0.4.2
pydantic==2.5.2
pydantic-settings==2.0.2
pytest==7.4.3
httpx==0.25.2
python-multipart==0.0.6

# dev
numpy==1.26.4
